+++ migrations/2019-11-04-150947_initial/down.sql
@@ -1,2 +1,0 @@
-DROP TABLE readings;
-DROP TABLE sensors;