+++ src/core/persistence/sensor.rs
@@ -1,0 +1,4 @@
+#[derive(PartialEq, Debug, Clone)]
+pub struct Sensor {
+    pub sensor: String,
+}