+++ src/core/persistence/thread.rs
@@ -1,7 +1,8 @@
+use crate::core::persistence::*;
 use crate::prelude::*;
 
 /// Spawn the persistence thread.
-pub fn spawn(db: Arc<Mutex<Db>>, tx: &Sender<Message>) -> Result<Sender<Message>> {
+pub fn spawn(db: Arc<Mutex<Connection>>, tx: &Sender<Message>) -> Result<Sender<Message>> {
     info!("Spawning readings persistence…");
     let tx = tx.clone();
     let (out_tx, rx) = crossbeam_channel::unbounded::<Message>();
@@ -17,34 +18,37 @@ pub fn spawn(db: Arc<Mutex<Db>>, tx: &Sender<Message>) -> Result<Sender<Message>
 }
 
 /// Process a message.
-fn process_message(message: Message, db: &Arc<Mutex<Db>>, tx: &Sender<Message>) -> Result<()> {
-    info!("{}: {:?} {:?}", &message.sensor, &message.type_, &message.value,);
+fn process_message(message: Message, db: &Arc<Mutex<Connection>>, tx: &Sender<Message>) -> Result<()> {
+    info!(
+        "{}: {:?} {:?}",
+        &message.sensor.sensor, &message.type_, &message.reading.value
+    );
     debug!("{:?}", &message);
     // TODO: handle `ReadSnapshot`.
     if message.type_ == MessageType::ReadLogged {
         let db = db.lock().unwrap();
-        let previous_reading = db.select_last_reading(&message.sensor)?;
-        db.upsert_reading(&message)?;
+        let previous_reading = select_last_reading(&db, &message.sensor.sensor)?;
+        upsert_reading(&db.lock().unwrap(), &message.sensor, &message.reading)?;
         send_messages(&previous_reading, &message, &tx)?;
     }
     Ok(())
 }
 
 /// Check if sensor value has been updated or changed and send corresponding messages.
-fn send_messages(previous_reading: &Option<Message>, message: &Message, tx: &Sender<Message>) -> Result<()> {
+fn send_messages(previous_reading: &Option<Reading>, message: &Message, tx: &Sender<Message>) -> Result<()> {
     if let Some(existing) = previous_reading {
-        if message.timestamp > existing.timestamp {
+        if message.reading.timestamp > existing.timestamp {
             tx.send(
-                Composer::new(format!("{}::update", &message.sensor))
+                Composer::new(format!("{}::update", &message.sensor.sensor))
                     .type_(MessageType::ReadNonLogged)
-                    .value(message.value.clone())
+                    .value(message.reading.value.clone())
                     .into(),
             )?;
-            if message.value != existing.value {
+            if message.reading.value != existing.value {
                 tx.send(
-                    Composer::new(format!("{}::change", &message.sensor))
+                    Composer::new(format!("{}::change", &message.sensor.sensor))
                         .type_(MessageType::ReadNonLogged)
-                        .value(message.value.clone())
+                        .value(message.reading.value.clone())
                         .into(),
                 )?;
             }